requirements.txt
python-dotenv==0.15.0
mysql-connector-python==8.0.23
pyserial==3.5